extract test classes for cooling schedules
