+++ src/test/java/org/cicirello/search/evo/BoltzmannSelectionTests.java
@@ -1,6 +1,6 @@
 /*
  * Chips-n-Salsa: A library of parallel self-adaptive local search algorithms.
- * Copyright (C) 2002-2022 Vincent A. Cicirello
+ * Copyright (C) 2002-2023 Vincent A. Cicirello
  *
  * This file is part of Chips-n-Salsa (https://chips-n-salsa.cicirello.org/).
  *
@@ -26,223 +26,6 @@ import org.junit.jupiter.api.*;
 
 /** JUnit test cases for Boltzmann selection and variations. */
 public class BoltzmannSelectionTests {
-
-  // Tests for the bias functions
-
-  @Test
-  public void testConstantBoltzmannBiasFunction() {
-    ConstantBoltzmannBiasFunction bias = new ConstantBoltzmannBiasFunction(1.0);
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(f), bias.bias(f));
-      bias.update();
-    }
-    bias.init();
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(f), bias.bias(f));
-      bias.update();
-    }
-
-    bias = new ConstantBoltzmannBiasFunction(0.5);
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(2 * f), bias.bias(f));
-      bias.update();
-    }
-    bias.init();
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(2 * f), bias.bias(f));
-      bias.update();
-    }
-
-    bias = new ConstantBoltzmannBiasFunction(2.0);
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(0.5 * f), bias.bias(f));
-      bias.update();
-    }
-    bias.init();
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(0.5 * f), bias.bias(f));
-      bias.update();
-    }
-
-    assertSame(bias, bias.split());
-  }
-
-  @Test
-  public void testLinearCoolingSchedule() {
-    LinearCoolingBiasFunction bias = new LinearCoolingBiasFunction(4.0, 1.0, 0.1);
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(0.25 * f), bias.bias(f));
-    }
-    bias.update();
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(f / 3.0), bias.bias(f));
-    }
-    bias.update();
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(0.5 * f), bias.bias(f));
-    }
-    bias.update();
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(f), bias.bias(f));
-    }
-    LinearCoolingBiasFunction biasSplit = bias.split();
-    bias.update();
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(10 * f), bias.bias(f));
-    }
-    bias.update();
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(10 * f), bias.bias(f));
-    }
-
-    // Use split copy
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(0.25 * f), biasSplit.bias(f));
-    }
-    biasSplit.update();
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(f / 3.0), biasSplit.bias(f));
-    }
-    biasSplit.update();
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(0.5 * f), biasSplit.bias(f));
-    }
-    biasSplit.update();
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(f), biasSplit.bias(f));
-    }
-    biasSplit.update();
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(10 * f), biasSplit.bias(f));
-    }
-    biasSplit.update();
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(10 * f), biasSplit.bias(f));
-    }
-
-    // reinitialize and use original
-    bias.init();
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(0.25 * f), bias.bias(f));
-    }
-    bias.update();
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(f / 3.0), bias.bias(f));
-    }
-    bias.update();
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(0.5 * f), bias.bias(f));
-    }
-    bias.update();
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(f), bias.bias(f));
-    }
-    bias.update();
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(10 * f), bias.bias(f));
-    }
-    bias.update();
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(10 * f), bias.bias(f));
-    }
-  }
-
-  @Test
-  public void testExponentialCoolingSchedule() {
-    ExponentialCoolingBiasFunction bias = new ExponentialCoolingBiasFunction(1.0, 0.5, 0.1);
-    double mult = 1.0;
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(mult * f), bias.bias(f));
-    }
-    bias.update();
-    mult *= 2.0;
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(mult * f), bias.bias(f));
-    }
-    bias.update();
-    mult *= 2.0;
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(mult * f), bias.bias(f));
-    }
-    bias.update();
-    mult *= 2.0;
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(mult * f), bias.bias(f));
-    }
-    ExponentialCoolingBiasFunction biasSplit = bias.split();
-    bias.update();
-    mult = 10;
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(mult * f), bias.bias(f));
-    }
-    bias.update();
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(mult * f), bias.bias(f));
-    }
-
-    // Use split copy
-    mult = 1.0;
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(mult * f), biasSplit.bias(f));
-    }
-    biasSplit.update();
-    mult *= 2.0;
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(mult * f), biasSplit.bias(f));
-    }
-    biasSplit.update();
-    mult *= 2.0;
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(mult * f), biasSplit.bias(f));
-    }
-    biasSplit.update();
-    mult *= 2.0;
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(mult * f), biasSplit.bias(f));
-    }
-    biasSplit.update();
-    mult = 10;
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(mult * f), biasSplit.bias(f));
-    }
-    biasSplit.update();
-    mult = 10;
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(mult * f), biasSplit.bias(f));
-    }
-
-    // reinitialize and use original
-    bias.init();
-    mult = 1.0;
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(mult * f), bias.bias(f));
-    }
-    bias.update();
-    mult *= 2.0;
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(mult * f), bias.bias(f));
-    }
-    bias.update();
-    mult *= 2.0;
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(mult * f), bias.bias(f));
-    }
-    bias.update();
-    mult *= 2.0;
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(mult * f), bias.bias(f));
-    }
-    bias.update();
-    mult = 10;
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(mult * f), bias.bias(f));
-    }
-    bias.update();
-    mult = 10;
-    for (double f = 0.0; f <= 5.1; f += 1.0) {
-      assertEquals(Math.exp(mult * f), bias.bias(f));
-    }
-  }
 
   // TESTS FOR THE STANDARD FITNESS PROPORTIONAL VERSION
